pyproject.toml
[project]
name = "qt-api"
version = "0.1"
description = "Python wrapper for Questrade API"
readme = "README.md"
requires-python = ">=3.8"
authors = [{name = "Lyubomir Georgiev"}]
license = {text = "Apache-2.0"}
classifiers = [
    "License :: OSI Approved :: Apache Software License"
]
dependencies = [
    "httpx",
    "pydantic",
    "PyYAML"
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project.urls]
Homepage = "https://github.com/LVG77/qt-api"
Changelog = "https://github.com/LVG77/qt-api/releases"
Issues = "https://github.com/LVG77/qt-api/issues"
CI = "https://github.com/LVG77/qt-api/actions"


[project.optional-dependencies]
test = ["pytest"]
dev  = ["pytest", "ipykernel"]